+++ pkgs/path/lib/src/characters.dart
@@ -5,6 +5,7 @@
 // Character-code constants.
 
 const hash = 0x23;
+const percent = 0x25;
 const plus = 0x2b;
 const minus = 0x2d;
 const period = 0x2e;
@@ -16,9 +17,9 @@ const question = 0x3f;
 const upperA = 0x41;
 const upperZ = 0x5a;
 const lowerA = 0x61;
+const lowerE = 0x65;
+const lowerF = 0x66;
+const lowerI = 0x69;
+const lowerL = 0x6c;
 const lowerZ = 0x7a;
 const backslash = 0x5c;
-
-bool isLetter(int char) => lowerA <= (char |= 0x20) && char <= lowerZ;
-
-bool isDigit(int char) => char ^ zero <= 9;